feat: Add functionality to save clip as asset
Introduces a new action to the store and a corresponding UI element in the context menu that allows users to save a selected clip as a new asset. This is useful for reusing portions of existing media or creating new assets from video segments. Currently, text clips are not supported for this operation.